requirements.txt
#
streamlit
qrcode

pillow


# pip install -r requirements.txt